# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: xg2g API
  description: |
    API for xg2g (Enigma2 to Plex/Jellyfin Proxy).

    **Authentication**:
    All endpoints support both **Bearer Token** (`Authorization: Bearer <token>`) and **Cookie** (`xg2g_session=<token>`) authentication.
    Streaming and media requests specifically support unified authentication since v3.1.5.

    **License**: PolyForm Noncommercial License 1.0.0.
    Commercial usage is restricted. See repository documentation for details.
  version: 3.0.0
servers:
  - url: /api/v3
    description: Production API
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    NowNextRequest:
      type: object
      required: [services]
      properties:
        services:
          type: array
          minItems: 1
          items:
            type: string
    NowNextEntry:
      type: object
      required: [title, start, end]
      properties:
        title:
          type: string
        start:
          type: integer
          description: Unix timestamp (seconds)
        end:
          type: integer
          description: Unix timestamp (seconds)
        startXmltv:
          type: string
          description: Original XMLTV start timestamp including offset
        endXmltv:
          type: string
          description: Original XMLTV end timestamp including offset
    NowNextItem:
      type: object
      required: [serviceRef]
      properties:
        serviceRef:
          type: string
        now:
          $ref: "#/components/schemas/NowNextEntry"
        next:
          $ref: "#/components/schemas/NowNextEntry"
    NowNextResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NowNextItem"
...
security:
  - BearerAuth: [v3:read]
paths:
...
  /services/now-next:
    post:
      summary: Get now/next EPG for a list of services
      operationId: postServicesNowNext
      tags:
        - services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NowNextRequest"
      responses:
        "200":
          description: Now/next EPG entries per service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NowNextResponse"
        "400":
          description: Invalid request
...